👌 rename `*.utils` in `*-utils`
