+++ scripts/release/update-peer-dependency-versions.js
@@ -1,6 +1,6 @@
 const lernaConfig = require('../../lerna.json')
-const { runMain } = require('../lib/execution.utils')
-const { modifyFile } = require('../lib/files.utils')
+const { runMain } = require('../lib/execution-utils')
+const { modifyFile } = require('../lib/files-utils')
 
 // This script updates the peer dependency versions between rum and logs packages to match the new
 // version during a release.
