+++ scripts/staging-ci/check-staging-merge.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const { command, printLog, printError, runMain } = require('../lib/execution.utils')
-const { initGitConfig } = require('../lib/git.utils')
+const { command, printLog, printError, runMain } = require('../lib/execution-utils')
+const { initGitConfig } = require('../lib/git-utils')
 
 const REPOSITORY = process.env.GIT_REPOSITORY
 const CI_COMMIT_SHA = process.env.CI_COMMIT_SHA
